pyproject.toml
[project]
name = "codealign"
version = "0.1.0"
description = "Code alignment and specification guard tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "streamlit",
    "groq",
    "google-generativeai",
    "numpy",
    "scikit-learn",
    "httpx",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
